requirements.txt
nextcord==2.0.0a10
requests == 2.27.1
SQLAlchemy==1.4.31
pygithub
python-dotenv
